Fix unit tests and make them take crop as float

+++ dali/pipeline/operators/resize/new_resize_test.cc
@@ -29,7 +29,7 @@ typedef enum {
 TYPED_TEST(NewResizeTest, TestFixedResizeAndCrop) {
   this->TstBody(this->DefaultSchema("ResizeCropMirror")
                   .AddArg("resize_shorter", 480.f)
-                  .AddArg("crop", vector<int>{224, 224}),
+                  .AddArg("crop", vector<float>{224, 224}),
                 this->getEps(t_TestFixedResizeAndCrop));
 }
 
@@ -37,7 +37,7 @@ TYPED_TEST(NewResizeTest, TestFixedResizeAndCropWarp) {
   this->TstBody(this->DefaultSchema("ResizeCropMirror")
                   .AddArg("resize_x", 480.f)
                   .AddArg("resize_y", 480.f)
-                  .AddArg("crop", vector<int>{224, 224}),
+                  .AddArg("crop", vector<float>{224, 224}),
                 this->getEps(t_TestFixedResizeAndCropWarp));
 }
